initial pass at unifying the codebases and using platformio for building
