+++ .gitignore
@@ -1,0 +1,3 @@
+.pioenvs
+.clang_complete
+.gcc-flags.json
