+++ lib/GoosciBleGatt/making_science_sensor_gatt.xml
