+++ lib/goosci/arduino_nrf_pins.h
