+++ lib/goosci/config_change.cpp
