+++ lib/goosci/goosci_utility.h
@@ -13,32 +13,22 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-#ifndef _DEBUG_PRINT_H_
+#ifndef _GOOSCI_UTILITY_H_
+#define _GOOSCI_UTILITY_H_
 
-#define DEBUG_PRINT(x) \
-  {                    \
-    if (Serial) {      \
-      Serial.print(x); \
-    }                  \
-  }
-#define DEBUG_PRINTLN(x) \
-  {                      \
-    if (Serial) {        \
-      Serial.println(x); \
-    }                    \
-  }
+#if defined(__ARDUINO_ARC__)
+  #include <CurieBLE.h>
+  void send_data(BLECharacteristic& characteristic, unsigned long timestamp, int value);
+#endif
 
-#define DEBUG_PRINT2(x, y) \
-  {                        \
-    if (Serial) {          \
-      Serial.print(x, y);  \
-    }                      \
-  }
-#define DEBUG_PRINTLN2(x, y) \
-  {                          \
-    if (Serial) {            \
-      Serial.println(x, y);  \
-    }                        \
-  }
+#if defined(__AVR_ATmega32U4__)
+  #include "GoosciBleGatt.h"
+  void send_data(GoosciBleGatt &goosciBle, unsigned long timestamp, int value);
+#endif
+
+// The serial port on leonardo will hang (UART buffer full) if the
+// serial port is not physically opened on the USB host.  This
+// function will block until the serial port is open.
+void wait_for_serial(void);
 
 #endif
