+++ lib/goosci/sensor.pb.c
