+++ lib/goosci/services.h
