+++ platformio.ini
@@ -1,0 +1,36 @@
+#
+# Project Configuration File
+#
+# A detailed documentation with the EXAMPLES is located here:
+# http://docs.platformio.org/en/latest/projectconf.html
+#
+
+# A sign `#` at the beginning of the line indicates a comment
+# Comment lines are ignored.
+
+# Simple and base environment
+# [env:mybaseenv]
+# platform = %INSTALLED_PLATFORM_NAME_HERE%
+# framework =
+# board =
+#
+# Automatic targets - enable auto-uploading
+# targets = upload
+
+[env:blend]
+platform = atmelavr
+framework = arduino
+board = blend
+src_filter = +<blend/>
+lib_use = SPI,EEPROM
+lib_dfcyclic = True
+lib_install = 431
+
+
+[env:101]
+platform = intel_arc32
+framework = arduino
+board = genuino101
+src_filter = +<arduino101/>
+lib_install = 431
+lib_ignore = Nordic_nRF8001,GoosciBleGatt
