--- src/blend/science-journal-arduino/config_change.h
@@ -1,30 +1,0 @@
-/*
- *  Copyright 2016 Google Inc. All Rights Reserved.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-#ifndef _CONFIG_CHANGE_H_
-#define _CONFIG_CHANGE_H_
-
-#include <pb_decode.h>
-enum PinType {
-  ANALOG = 1,
-  DIGITAL = 2,
-  VIRTUAL = 3,
-};
-
-
-bool decode_pin(pb_istream_t *stream, const pb_field_t *field, void * *arg);
-void handle(uint8_t* data, int8_t length);
-
-#endif
