--- src/blend/science-journal-arduino/sensor.pb.c
@@ -1,226 +1,0 @@
-/* Automatically generated nanopb constant definitions */
-/* Generated by nanopb-0.3.5-dev at Tue Oct 27 09:54:57 2015. */
-
-#include "sensor.pb.h"
-
-#if PB_PROTO_HEADER_VERSION != 30
-#error Regenerate this file with the current version of nanopb generator.
-#endif
-
-const int32_t goosci_Version_version_default = 1;
-
-
-const pb_field_t goosci_Empty_fields[1] = {
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_Version_fields[2] = {
-    PB_FIELD(  1, INT32   , REQUIRED, STATIC  , FIRST, goosci_Version, version, version, &goosci_Version_version_default),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_Unit_fields[2] = {
-    PB_FIELD(  1, UENUM   , REQUIRED, STATIC  , FIRST, goosci_Unit, units, units, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_AnalogExtremum_fields[2] = {
-    PB_FIELD(  1, INT32   , REQUIRED, STATIC  , FIRST, goosci_AnalogExtremum, extremum, extremum, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_AnalogZero_fields[2] = {
-    PB_FIELD(  1, INT32   , REQUIRED, STATIC  , FIRST, goosci_AnalogZero, zero, zero, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_AnalogExtrema_fields[3] = {
-    PB_FIELD(  1, MESSAGE , OPTIONAL, STATIC  , FIRST, goosci_AnalogExtrema, minimum, minimum, &goosci_AnalogExtremum_fields),
-    PB_FIELD(  2, MESSAGE , OPTIONAL, STATIC  , OTHER, goosci_AnalogExtrema, maximum, minimum, &goosci_AnalogExtremum_fields),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_FloatExtremum_fields[2] = {
-    PB_FIELD(  1, INT32   , REQUIRED, STATIC  , FIRST, goosci_FloatExtremum, extremum, extremum, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_FloatZero_fields[2] = {
-    PB_FIELD(  1, INT32   , REQUIRED, STATIC  , FIRST, goosci_FloatZero, zero, zero, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_FloatExtrema_fields[3] = {
-    PB_FIELD(  1, MESSAGE , OPTIONAL, STATIC  , FIRST, goosci_FloatExtrema, minimum, minimum, &goosci_FloatExtremum_fields),
-    PB_FIELD(  2, MESSAGE , OPTIONAL, STATIC  , OTHER, goosci_FloatExtrema, maximum, minimum, &goosci_FloatExtremum_fields),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_IntExtremum_fields[2] = {
-    PB_FIELD(  1, INT32   , REQUIRED, STATIC  , FIRST, goosci_IntExtremum, extremum, extremum, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_IntZero_fields[2] = {
-    PB_FIELD(  1, INT32   , REQUIRED, STATIC  , FIRST, goosci_IntZero, zero, zero, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_IntExtrema_fields[3] = {
-    PB_FIELD(  1, MESSAGE , OPTIONAL, STATIC  , FIRST, goosci_IntExtrema, minimum, minimum, &goosci_IntExtremum_fields),
-    PB_FIELD(  2, MESSAGE , OPTIONAL, STATIC  , OTHER, goosci_IntExtrema, maximum, minimum, &goosci_IntExtremum_fields),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_Range_fields[4] = {
-    PB_ONEOF_FIELD(extrema,   1, MESSAGE , ONEOF, STATIC  , FIRST, goosci_Range, analog_extrema, analog_extrema, &goosci_AnalogExtrema_fields),
-    PB_ONEOF_FIELD(extrema,   2, MESSAGE , ONEOF, STATIC  , FIRST, goosci_Range, float_extrema, float_extrema, &goosci_FloatExtrema_fields),
-    PB_ONEOF_FIELD(extrema,   3, MESSAGE , ONEOF, STATIC  , FIRST, goosci_Range, int_extrema, int_extrema, &goosci_IntExtrema_fields),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_Zero_fields[4] = {
-    PB_ONEOF_FIELD(zero,   1, MESSAGE , ONEOF, STATIC  , FIRST, goosci_Zero, analog_zero, analog_zero, &goosci_AnalogZero_fields),
-    PB_ONEOF_FIELD(zero,   2, MESSAGE , ONEOF, STATIC  , FIRST, goosci_Zero, float_zero, float_zero, &goosci_FloatZero_fields),
-    PB_ONEOF_FIELD(zero,   3, MESSAGE , ONEOF, STATIC  , FIRST, goosci_Zero, int_zero, int_zero, &goosci_IntZero_fields),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_AxisConfig_fields[3] = {
-    PB_FIELD(  1, MESSAGE , OPTIONAL, STATIC  , FIRST, goosci_AxisConfig, range, range, &goosci_Range_fields),
-    PB_FIELD(  2, MESSAGE , OPTIONAL, STATIC  , OTHER, goosci_AxisConfig, zero_value, range, &goosci_Zero_fields),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_AnalogPin_fields[2] = {
-    PB_FIELD(  1, INT32   , REQUIRED, STATIC  , FIRST, goosci_AnalogPin, pin, pin, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_DigitalPin_fields[2] = {
-    PB_FIELD(  1, INT32   , REQUIRED, STATIC  , FIRST, goosci_DigitalPin, pin, pin, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_ScalarDescriptor_fields[5] = {
-    PB_FIELD(  1, STRING  , REQUIRED, CALLBACK, FIRST, goosci_ScalarDescriptor, label, label, 0),
-    PB_FIELD(  2, UENUM   , REQUIRED, STATIC  , OTHER, goosci_ScalarDescriptor, type, label, 0),
-    PB_FIELD(  3, MESSAGE , OPTIONAL, STATIC  , OTHER, goosci_ScalarDescriptor, axis_config, type, &goosci_AxisConfig_fields),
-    PB_FIELD(  4, MESSAGE , OPTIONAL, STATIC  , OTHER, goosci_ScalarDescriptor, unit, axis_config, &goosci_Unit_fields),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_VirtualPin_fields[2] = {
-    PB_FIELD(  1, INT32   , REQUIRED, STATIC  , FIRST, goosci_VirtualPin, pin, pin, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_Pin_fields[4] = {
-    PB_ONEOF_FIELD(pin,  10, MESSAGE , ONEOF, STATIC  , FIRST, goosci_Pin, analog_pin, analog_pin, &goosci_AnalogPin_fields),
-    PB_ONEOF_FIELD(pin,  11, MESSAGE , ONEOF, STATIC  , FIRST, goosci_Pin, digital_pin, digital_pin, &goosci_DigitalPin_fields),
-    PB_ONEOF_FIELD(pin,  12, MESSAGE , ONEOF, STATIC  , FIRST, goosci_Pin, virtual_pin, virtual_pin, &goosci_VirtualPin_fields),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_PinDescriptor_fields[4] = {
-    PB_FIELD(  1, STRING  , REQUIRED, CALLBACK, FIRST, goosci_PinDescriptor, label, label, 0),
-    PB_FIELD(  2, MESSAGE , REQUIRED, STATIC  , OTHER, goosci_PinDescriptor, pin, label, &goosci_Pin_fields),
-    PB_FIELD(  3, MESSAGE , REQUIRED, STATIC  , OTHER, goosci_PinDescriptor, scalar_descriptor, pin, &goosci_ScalarDescriptor_fields),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_Interval_fields[3] = {
-    PB_FIELD(  1, INT32   , REQUIRED, STATIC  , FIRST, goosci_Interval, count, count, 0),
-    PB_FIELD(  2, INT32   , OPTIONAL, STATIC  , OTHER, goosci_Interval, frequency, count, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_SensorDataRequest_fields[4] = {
-    PB_FIELD(  1, UINT32  , REQUIRED, STATIC  , FIRST, goosci_SensorDataRequest, timestamp_key, timestamp_key, 0),
-    PB_FIELD(  2, MESSAGE , REQUIRED, STATIC  , OTHER, goosci_SensorDataRequest, interval, timestamp_key, &goosci_Interval_fields),
-    PB_FIELD(  3, MESSAGE , REPEATED, CALLBACK, OTHER, goosci_SensorDataRequest, pin, interval, &goosci_Pin_fields),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_BleSensorConfig_fields[2] = {
-    PB_FIELD(  1, STRING  , OPTIONAL, CALLBACK, FIRST, goosci_BleSensorConfig, address, address, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_SensorEnumeration_fields[2] = {
-    PB_FIELD(  1, MESSAGE , REPEATED, CALLBACK, FIRST, goosci_SensorEnumeration, pin_descriptor, pin_descriptor, &goosci_PinDescriptor_fields),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_AnalogValue_fields[2] = {
-    PB_FIELD(  1, UINT32  , REQUIRED, STATIC  , FIRST, goosci_AnalogValue, value, value, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_DigitalValue_fields[2] = {
-    PB_FIELD(  1, BOOL    , REQUIRED, STATIC  , FIRST, goosci_DigitalValue, value, value, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_FloatValue_fields[2] = {
-    PB_FIELD(  1, FLOAT   , REQUIRED, STATIC  , FIRST, goosci_FloatValue, value, value, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_IntValue_fields[2] = {
-    PB_FIELD(  1, INT32   , REQUIRED, STATIC  , FIRST, goosci_IntValue, value, value, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_StringValue_fields[2] = {
-    PB_FIELD(  1, STRING  , REQUIRED, CALLBACK, FIRST, goosci_StringValue, value, value, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_Data_fields[7] = {
-    PB_FIELD(  1, MESSAGE , REQUIRED, STATIC  , FIRST, goosci_Data, pin, pin, &goosci_Pin_fields),
-    PB_ONEOF_FIELD(value,  10, MESSAGE , ONEOF, STATIC  , OTHER, goosci_Data, analog_value, pin, &goosci_AnalogValue_fields),
-    PB_ONEOF_FIELD(value,  11, MESSAGE , ONEOF, STATIC  , OTHER, goosci_Data, digital_value, pin, &goosci_DigitalValue_fields),
-    PB_ONEOF_FIELD(value,  12, MESSAGE , ONEOF, STATIC  , OTHER, goosci_Data, float_value, pin, &goosci_FloatValue_fields),
-    PB_ONEOF_FIELD(value,  13, MESSAGE , ONEOF, STATIC  , OTHER, goosci_Data, int_value, pin, &goosci_IntValue_fields),
-    PB_ONEOF_FIELD(value,  14, MESSAGE , ONEOF, STATIC  , OTHER, goosci_Data, string_value, pin, &goosci_StringValue_fields),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_Error_fields[2] = {
-    PB_FIELD(  1, STRING  , REQUIRED, CALLBACK, FIRST, goosci_Error, error, error, 0),
-    PB_LAST_FIELD
-};
-
-const pb_field_t goosci_SensorData_fields[4] = {
-    PB_FIELD(  1, UINT32  , REQUIRED, STATIC  , FIRST, goosci_SensorData, timestamp_key, timestamp_key, 0),
-    PB_ONEOF_FIELD(result,  10, MESSAGE , ONEOF, STATIC  , OTHER, goosci_SensorData, error, timestamp_key, &goosci_Error_fields),
-    PB_ONEOF_FIELD(result,  11, MESSAGE , ONEOF, STATIC  , OTHER, goosci_SensorData, data, timestamp_key, &goosci_Data_fields),
-    PB_LAST_FIELD
-};
-
-
-/* Check that field information fits in pb_field_t */
-#if !defined(PB_FIELD_32BIT)
-/* If you get an error here, it means that you need to define PB_FIELD_32BIT
- * compile-time option. You can do that in pb.h or on compiler command line.
- * 
- * The reason you need to do this is that some of your messages contain tag
- * numbers or field sizes that are larger than what can fit in 8 or 16 bit
- * field descriptors.
- */
-PB_STATIC_ASSERT((pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_Zero, zero.analog_zero) < 65536 && pb_membersize(goosci_Zero, zero.float_zero) < 65536 && pb_membersize(goosci_Zero, zero.int_zero) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_Zero, zero.analog_zero) < 65536 && pb_membersize(goosci_Zero, zero.float_zero) < 65536 && pb_membersize(goosci_Zero, zero.int_zero) < 65536 && pb_membersize(goosci_AxisConfig, range) < 65536 && pb_membersize(goosci_AxisConfig, zero_value) < 65536 && pb_membersize(goosci_ScalarDescriptor, axis_config) < 65536 && pb_membersize(goosci_ScalarDescriptor, unit) < 65536 && pb_membersize(goosci_Pin, pin.analog_pin) < 65536 && pb_membersize(goosci_Pin, pin.digital_pin) < 65536 && pb_membersize(goosci_Pin, pin.virtual_pin) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_Zero, zero.analog_zero) < 65536 && pb_membersize(goosci_Zero, zero.float_zero) < 65536 && pb_membersize(goosci_Zero, zero.int_zero) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_Zero, zero.analog_zero) < 65536 && pb_membersize(goosci_Zero, zero.float_zero) < 65536 && pb_membersize(goosci_Zero, zero.int_zero) < 65536 && pb_membersize(goosci_AxisConfig, range) < 65536 && pb_membersize(goosci_AxisConfig, zero_value) < 65536 && pb_membersize(goosci_ScalarDescriptor, axis_config) < 65536 && pb_membersize(goosci_ScalarDescriptor, unit) < 65536 && pb_membersize(goosci_Pin, pin.analog_pin) < 65536 && pb_membersize(goosci_Pin, pin.digital_pin) < 65536 && pb_membersize(goosci_Pin, pin.virtual_pin) < 65536 && pb_membersize(goosci_PinDescriptor, pin) < 65536 && pb_membersize(goosci_PinDescriptor, scalar_descriptor) < 65536 && pb_membersize(goosci_SensorDataRequest, interval) < 65536 && pb_membersize(goosci_SensorDataRequest, pin) < 65536 && pb_membersize(goosci_SensorEnumeration, pin_descriptor) < 65536 && pb_membersize(goosci_Data, value.analog_value) < 65536 && pb_membersize(goosci_Data, value.digital_value) < 65536 && pb_membersize(goosci_Data, value.float_value) < 65536 && pb_membersize(goosci_Data, value.int_value) < 65536 && pb_membersize(goosci_Data, value.string_value) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_Zero, zero.analog_zero) < 65536 && pb_membersize(goosci_Zero, zero.float_zero) < 65536 && pb_membersize(goosci_Zero, zero.int_zero) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_Zero, zero.analog_zero) < 65536 && pb_membersize(goosci_Zero, zero.float_zero) < 65536 && pb_membersize(goosci_Zero, zero.int_zero) < 65536 && pb_membersize(goosci_AxisConfig, range) < 65536 && pb_membersize(goosci_AxisConfig, zero_value) < 65536 && pb_membersize(goosci_ScalarDescriptor, axis_config) < 65536 && pb_membersize(goosci_ScalarDescriptor, unit) < 65536 && pb_membersize(goosci_Pin, pin.analog_pin) < 65536 && pb_membersize(goosci_Pin, pin.digital_pin) < 65536 && pb_membersize(goosci_Pin, pin.virtual_pin) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_Zero, zero.analog_zero) < 65536 && pb_membersize(goosci_Zero, zero.float_zero) < 65536 && pb_membersize(goosci_Zero, zero.int_zero) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_Zero, zero.analog_zero) < 65536 && pb_membersize(goosci_Zero, zero.float_zero) < 65536 && pb_membersize(goosci_Zero, zero.int_zero) < 65536 && pb_membersize(goosci_AxisConfig, range) < 65536 && pb_membersize(goosci_AxisConfig, zero_value) < 65536 && pb_membersize(goosci_ScalarDescriptor, axis_config) < 65536 && pb_membersize(goosci_ScalarDescriptor, unit) < 65536 && pb_membersize(goosci_Pin, pin.analog_pin) < 65536 && pb_membersize(goosci_Pin, pin.digital_pin) < 65536 && pb_membersize(goosci_Pin, pin.virtual_pin) < 65536 && pb_membersize(goosci_PinDescriptor, pin) < 65536 && pb_membersize(goosci_PinDescriptor, scalar_descriptor) < 65536 && pb_membersize(goosci_SensorDataRequest, interval) < 65536 && pb_membersize(goosci_SensorDataRequest, pin) < 65536 && pb_membersize(goosci_SensorEnumeration, pin_descriptor) < 65536 && pb_membersize(goosci_Data, value.analog_value) < 65536 && pb_membersize(goosci_Data, value.digital_value) < 65536 && pb_membersize(goosci_Data, value.float_value) < 65536 && pb_membersize(goosci_Data, value.int_value) < 65536 && pb_membersize(goosci_Data, value.string_value) < 65536 && pb_membersize(goosci_Data, pin) < 65536 && pb_membersize(goosci_SensorData, result.error) < 65536 && pb_membersize(goosci_SensorData, result.data) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_Zero, zero.analog_zero) < 65536 && pb_membersize(goosci_Zero, zero.float_zero) < 65536 && pb_membersize(goosci_Zero, zero.int_zero) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_Zero, zero.analog_zero) < 65536 && pb_membersize(goosci_Zero, zero.float_zero) < 65536 && pb_membersize(goosci_Zero, zero.int_zero) < 65536 && pb_membersize(goosci_AxisConfig, range) < 65536 && pb_membersize(goosci_AxisConfig, zero_value) < 65536 && pb_membersize(goosci_ScalarDescriptor, axis_config) < 65536 && pb_membersize(goosci_ScalarDescriptor, unit) < 65536 && pb_membersize(goosci_Pin, pin.analog_pin) < 65536 && pb_membersize(goosci_Pin, pin.digital_pin) < 65536 && pb_membersize(goosci_Pin, pin.virtual_pin) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_Zero, zero.analog_zero) < 65536 && pb_membersize(goosci_Zero, zero.float_zero) < 65536 && pb_membersize(goosci_Zero, zero.int_zero) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_Zero, zero.analog_zero) < 65536 && pb_membersize(goosci_Zero, zero.float_zero) < 65536 && pb_membersize(goosci_Zero, zero.int_zero) < 65536 && pb_membersize(goosci_AxisConfig, range) < 65536 && pb_membersize(goosci_AxisConfig, zero_value) < 65536 && pb_membersize(goosci_ScalarDescriptor, axis_config) < 65536 && pb_membersize(goosci_ScalarDescriptor, unit) < 65536 && pb_membersize(goosci_Pin, pin.analog_pin) < 65536 && pb_membersize(goosci_Pin, pin.digital_pin) < 65536 && pb_membersize(goosci_Pin, pin.virtual_pin) < 65536 && pb_membersize(goosci_PinDescriptor, pin) < 65536 && pb_membersize(goosci_PinDescriptor, scalar_descriptor) < 65536 && pb_membersize(goosci_SensorDataRequest, interval) < 65536 && pb_membersize(goosci_SensorDataRequest, pin) < 65536 && pb_membersize(goosci_SensorEnumeration, pin_descriptor) < 65536 && pb_membersize(goosci_Data, value.analog_value) < 65536 && pb_membersize(goosci_Data, value.digital_value) < 65536 && pb_membersize(goosci_Data, value.float_value) < 65536 && pb_membersize(goosci_Data, value.int_value) < 65536 && pb_membersize(goosci_Data, value.string_value) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_Zero, zero.analog_zero) < 65536 && pb_membersize(goosci_Zero, zero.float_zero) < 65536 && pb_membersize(goosci_Zero, zero.int_zero) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_Zero, zero.analog_zero) < 65536 && pb_membersize(goosci_Zero, zero.float_zero) < 65536 && pb_membersize(goosci_Zero, zero.int_zero) < 65536 && pb_membersize(goosci_AxisConfig, range) < 65536 && pb_membersize(goosci_AxisConfig, zero_value) < 65536 && pb_membersize(goosci_ScalarDescriptor, axis_config) < 65536 && pb_membersize(goosci_ScalarDescriptor, unit) < 65536 && pb_membersize(goosci_Pin, pin.analog_pin) < 65536 && pb_membersize(goosci_Pin, pin.digital_pin) < 65536 && pb_membersize(goosci_Pin, pin.virtual_pin) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_Zero, zero.analog_zero) < 65536 && pb_membersize(goosci_Zero, zero.float_zero) < 65536 && pb_membersize(goosci_Zero, zero.int_zero) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_AnalogExtrema, minimum) < 65536 && pb_membersize(goosci_AnalogExtrema, maximum) < 65536 && pb_membersize(goosci_FloatExtrema, minimum) < 65536 && pb_membersize(goosci_FloatExtrema, maximum) < 65536 && pb_membersize(goosci_IntExtrema, minimum) < 65536 && pb_membersize(goosci_IntExtrema, maximum) < 65536 && pb_membersize(goosci_Range, extrema.analog_extrema) < 65536 && pb_membersize(goosci_Range, extrema.float_extrema) < 65536 && pb_membersize(goosci_Range, extrema.int_extrema) < 65536 && pb_membersize(goosci_Zero, zero.analog_zero) < 65536 && pb_membersize(goosci_Zero, zero.float_zero) < 65536 && pb_membersize(goosci_Zero, zero.int_zero) < 65536 && pb_membersize(goosci_AxisConfig, range) < 65536 && pb_membersize(goosci_AxisConfig, zero_value) < 65536 && pb_membersize(goosci_ScalarDescriptor, axis_config) < 65536 && pb_membersize(goosci_ScalarDescriptor, unit) < 65536 && pb_membersize(goosci_Pin, pin.analog_pin) < 65536 && pb_membersize(goosci_Pin, pin.digital_pin) < 65536 && pb_membersize(goosci_Pin, pin.virtual_pin) < 65536 && pb_membersize(goosci_PinDescriptor, pin) < 65536 && pb_membersize(goosci_PinDescriptor, scalar_descriptor) < 65536 && pb_membersize(goosci_SensorDataRequest, interval) < 65536 && pb_membersize(goosci_SensorDataRequest, pin) < 65536 && pb_membersize(goosci_SensorEnumeration, pin_descriptor) < 65536 && pb_membersize(goosci_Data, value.analog_value) < 65536 && pb_membersize(goosci_Data, value.digital_value) < 65536 && pb_membersize(goosci_Data, value.float_value) < 65536 && pb_membersize(goosci_Data, value.int_value) < 65536 && pb_membersize(goosci_Data, value.string_value) < 65536 && pb_membersize(goosci_Data, pin) < 65536 && pb_membersize(goosci_SensorData, result.error) < 65536 && pb_membersize(goosci_SensorData, result.data) < 65536), YOU_MUST_DEFINE_PB_FIELD_32BIT_FOR_MESSAGES_goosci_Empty_goosci_Version_goosci_Unit_goosci_AnalogExtremum_goosci_AnalogZero_goosci_AnalogExtrema_goosci_FloatExtremum_goosci_FloatZero_goosci_FloatExtrema_goosci_IntExtremum_goosci_IntZero_goosci_IntExtrema_goosci_Range_goosci_Zero_goosci_AxisConfig_goosci_AnalogPin_goosci_DigitalPin_goosci_ScalarDescriptor_goosci_VirtualPin_goosci_Pin_goosci_PinDescriptor_goosci_Interval_goosci_SensorDataRequest_goosci_BleSensorConfig_goosci_SensorEnumeration_goosci_AnalogValue_goosci_DigitalValue_goosci_FloatValue_goosci_IntValue_goosci_StringValue_goosci_Data_goosci_Error_goosci_SensorData)
-#endif
-
-#if !defined(PB_FIELD_16BIT) && !defined(PB_FIELD_32BIT)
-/* If you get an error here, it means that you need to define PB_FIELD_16BIT
- * compile-time option. You can do that in pb.h or on compiler command line.
- * 
- * The reason you need to do this is that some of your messages contain tag
- * numbers or field sizes that are larger than what can fit in the default
- * 8 bit descriptors.
- */
-PB_STATIC_ASSERT((pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_Zero, zero.analog_zero) < 256 && pb_membersize(goosci_Zero, zero.float_zero) < 256 && pb_membersize(goosci_Zero, zero.int_zero) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_Zero, zero.analog_zero) < 256 && pb_membersize(goosci_Zero, zero.float_zero) < 256 && pb_membersize(goosci_Zero, zero.int_zero) < 256 && pb_membersize(goosci_AxisConfig, range) < 256 && pb_membersize(goosci_AxisConfig, zero_value) < 256 && pb_membersize(goosci_ScalarDescriptor, axis_config) < 256 && pb_membersize(goosci_ScalarDescriptor, unit) < 256 && pb_membersize(goosci_Pin, pin.analog_pin) < 256 && pb_membersize(goosci_Pin, pin.digital_pin) < 256 && pb_membersize(goosci_Pin, pin.virtual_pin) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_Zero, zero.analog_zero) < 256 && pb_membersize(goosci_Zero, zero.float_zero) < 256 && pb_membersize(goosci_Zero, zero.int_zero) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_Zero, zero.analog_zero) < 256 && pb_membersize(goosci_Zero, zero.float_zero) < 256 && pb_membersize(goosci_Zero, zero.int_zero) < 256 && pb_membersize(goosci_AxisConfig, range) < 256 && pb_membersize(goosci_AxisConfig, zero_value) < 256 && pb_membersize(goosci_ScalarDescriptor, axis_config) < 256 && pb_membersize(goosci_ScalarDescriptor, unit) < 256 && pb_membersize(goosci_Pin, pin.analog_pin) < 256 && pb_membersize(goosci_Pin, pin.digital_pin) < 256 && pb_membersize(goosci_Pin, pin.virtual_pin) < 256 && pb_membersize(goosci_PinDescriptor, pin) < 256 && pb_membersize(goosci_PinDescriptor, scalar_descriptor) < 256 && pb_membersize(goosci_SensorDataRequest, interval) < 256 && pb_membersize(goosci_SensorDataRequest, pin) < 256 && pb_membersize(goosci_SensorEnumeration, pin_descriptor) < 256 && pb_membersize(goosci_Data, value.analog_value) < 256 && pb_membersize(goosci_Data, value.digital_value) < 256 && pb_membersize(goosci_Data, value.float_value) < 256 && pb_membersize(goosci_Data, value.int_value) < 256 && pb_membersize(goosci_Data, value.string_value) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_Zero, zero.analog_zero) < 256 && pb_membersize(goosci_Zero, zero.float_zero) < 256 && pb_membersize(goosci_Zero, zero.int_zero) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_Zero, zero.analog_zero) < 256 && pb_membersize(goosci_Zero, zero.float_zero) < 256 && pb_membersize(goosci_Zero, zero.int_zero) < 256 && pb_membersize(goosci_AxisConfig, range) < 256 && pb_membersize(goosci_AxisConfig, zero_value) < 256 && pb_membersize(goosci_ScalarDescriptor, axis_config) < 256 && pb_membersize(goosci_ScalarDescriptor, unit) < 256 && pb_membersize(goosci_Pin, pin.analog_pin) < 256 && pb_membersize(goosci_Pin, pin.digital_pin) < 256 && pb_membersize(goosci_Pin, pin.virtual_pin) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_Zero, zero.analog_zero) < 256 && pb_membersize(goosci_Zero, zero.float_zero) < 256 && pb_membersize(goosci_Zero, zero.int_zero) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_Zero, zero.analog_zero) < 256 && pb_membersize(goosci_Zero, zero.float_zero) < 256 && pb_membersize(goosci_Zero, zero.int_zero) < 256 && pb_membersize(goosci_AxisConfig, range) < 256 && pb_membersize(goosci_AxisConfig, zero_value) < 256 && pb_membersize(goosci_ScalarDescriptor, axis_config) < 256 && pb_membersize(goosci_ScalarDescriptor, unit) < 256 && pb_membersize(goosci_Pin, pin.analog_pin) < 256 && pb_membersize(goosci_Pin, pin.digital_pin) < 256 && pb_membersize(goosci_Pin, pin.virtual_pin) < 256 && pb_membersize(goosci_PinDescriptor, pin) < 256 && pb_membersize(goosci_PinDescriptor, scalar_descriptor) < 256 && pb_membersize(goosci_SensorDataRequest, interval) < 256 && pb_membersize(goosci_SensorDataRequest, pin) < 256 && pb_membersize(goosci_SensorEnumeration, pin_descriptor) < 256 && pb_membersize(goosci_Data, value.analog_value) < 256 && pb_membersize(goosci_Data, value.digital_value) < 256 && pb_membersize(goosci_Data, value.float_value) < 256 && pb_membersize(goosci_Data, value.int_value) < 256 && pb_membersize(goosci_Data, value.string_value) < 256 && pb_membersize(goosci_Data, pin) < 256 && pb_membersize(goosci_SensorData, result.error) < 256 && pb_membersize(goosci_SensorData, result.data) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_Zero, zero.analog_zero) < 256 && pb_membersize(goosci_Zero, zero.float_zero) < 256 && pb_membersize(goosci_Zero, zero.int_zero) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_Zero, zero.analog_zero) < 256 && pb_membersize(goosci_Zero, zero.float_zero) < 256 && pb_membersize(goosci_Zero, zero.int_zero) < 256 && pb_membersize(goosci_AxisConfig, range) < 256 && pb_membersize(goosci_AxisConfig, zero_value) < 256 && pb_membersize(goosci_ScalarDescriptor, axis_config) < 256 && pb_membersize(goosci_ScalarDescriptor, unit) < 256 && pb_membersize(goosci_Pin, pin.analog_pin) < 256 && pb_membersize(goosci_Pin, pin.digital_pin) < 256 && pb_membersize(goosci_Pin, pin.virtual_pin) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_Zero, zero.analog_zero) < 256 && pb_membersize(goosci_Zero, zero.float_zero) < 256 && pb_membersize(goosci_Zero, zero.int_zero) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_Zero, zero.analog_zero) < 256 && pb_membersize(goosci_Zero, zero.float_zero) < 256 && pb_membersize(goosci_Zero, zero.int_zero) < 256 && pb_membersize(goosci_AxisConfig, range) < 256 && pb_membersize(goosci_AxisConfig, zero_value) < 256 && pb_membersize(goosci_ScalarDescriptor, axis_config) < 256 && pb_membersize(goosci_ScalarDescriptor, unit) < 256 && pb_membersize(goosci_Pin, pin.analog_pin) < 256 && pb_membersize(goosci_Pin, pin.digital_pin) < 256 && pb_membersize(goosci_Pin, pin.virtual_pin) < 256 && pb_membersize(goosci_PinDescriptor, pin) < 256 && pb_membersize(goosci_PinDescriptor, scalar_descriptor) < 256 && pb_membersize(goosci_SensorDataRequest, interval) < 256 && pb_membersize(goosci_SensorDataRequest, pin) < 256 && pb_membersize(goosci_SensorEnumeration, pin_descriptor) < 256 && pb_membersize(goosci_Data, value.analog_value) < 256 && pb_membersize(goosci_Data, value.digital_value) < 256 && pb_membersize(goosci_Data, value.float_value) < 256 && pb_membersize(goosci_Data, value.int_value) < 256 && pb_membersize(goosci_Data, value.string_value) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_Zero, zero.analog_zero) < 256 && pb_membersize(goosci_Zero, zero.float_zero) < 256 && pb_membersize(goosci_Zero, zero.int_zero) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_Zero, zero.analog_zero) < 256 && pb_membersize(goosci_Zero, zero.float_zero) < 256 && pb_membersize(goosci_Zero, zero.int_zero) < 256 && pb_membersize(goosci_AxisConfig, range) < 256 && pb_membersize(goosci_AxisConfig, zero_value) < 256 && pb_membersize(goosci_ScalarDescriptor, axis_config) < 256 && pb_membersize(goosci_ScalarDescriptor, unit) < 256 && pb_membersize(goosci_Pin, pin.analog_pin) < 256 && pb_membersize(goosci_Pin, pin.digital_pin) < 256 && pb_membersize(goosci_Pin, pin.virtual_pin) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_Zero, zero.analog_zero) < 256 && pb_membersize(goosci_Zero, zero.float_zero) < 256 && pb_membersize(goosci_Zero, zero.int_zero) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_AnalogExtrema, minimum) < 256 && pb_membersize(goosci_AnalogExtrema, maximum) < 256 && pb_membersize(goosci_FloatExtrema, minimum) < 256 && pb_membersize(goosci_FloatExtrema, maximum) < 256 && pb_membersize(goosci_IntExtrema, minimum) < 256 && pb_membersize(goosci_IntExtrema, maximum) < 256 && pb_membersize(goosci_Range, extrema.analog_extrema) < 256 && pb_membersize(goosci_Range, extrema.float_extrema) < 256 && pb_membersize(goosci_Range, extrema.int_extrema) < 256 && pb_membersize(goosci_Zero, zero.analog_zero) < 256 && pb_membersize(goosci_Zero, zero.float_zero) < 256 && pb_membersize(goosci_Zero, zero.int_zero) < 256 && pb_membersize(goosci_AxisConfig, range) < 256 && pb_membersize(goosci_AxisConfig, zero_value) < 256 && pb_membersize(goosci_ScalarDescriptor, axis_config) < 256 && pb_membersize(goosci_ScalarDescriptor, unit) < 256 && pb_membersize(goosci_Pin, pin.analog_pin) < 256 && pb_membersize(goosci_Pin, pin.digital_pin) < 256 && pb_membersize(goosci_Pin, pin.virtual_pin) < 256 && pb_membersize(goosci_PinDescriptor, pin) < 256 && pb_membersize(goosci_PinDescriptor, scalar_descriptor) < 256 && pb_membersize(goosci_SensorDataRequest, interval) < 256 && pb_membersize(goosci_SensorDataRequest, pin) < 256 && pb_membersize(goosci_SensorEnumeration, pin_descriptor) < 256 && pb_membersize(goosci_Data, value.analog_value) < 256 && pb_membersize(goosci_Data, value.digital_value) < 256 && pb_membersize(goosci_Data, value.float_value) < 256 && pb_membersize(goosci_Data, value.int_value) < 256 && pb_membersize(goosci_Data, value.string_value) < 256 && pb_membersize(goosci_Data, pin) < 256 && pb_membersize(goosci_SensorData, result.error) < 256 && pb_membersize(goosci_SensorData, result.data) < 256), YOU_MUST_DEFINE_PB_FIELD_16BIT_FOR_MESSAGES_goosci_Empty_goosci_Version_goosci_Unit_goosci_AnalogExtremum_goosci_AnalogZero_goosci_AnalogExtrema_goosci_FloatExtremum_goosci_FloatZero_goosci_FloatExtrema_goosci_IntExtremum_goosci_IntZero_goosci_IntExtrema_goosci_Range_goosci_Zero_goosci_AxisConfig_goosci_AnalogPin_goosci_DigitalPin_goosci_ScalarDescriptor_goosci_VirtualPin_goosci_Pin_goosci_PinDescriptor_goosci_Interval_goosci_SensorDataRequest_goosci_BleSensorConfig_goosci_SensorEnumeration_goosci_AnalogValue_goosci_DigitalValue_goosci_FloatValue_goosci_IntValue_goosci_StringValue_goosci_Data_goosci_Error_goosci_SensorData)
-#endif
-
-
